Most of User Story 11: Pressure Heatmap
PatientSessionData and PatientSnapshotData models handle the pressure data from sessions.
CSV files are read if they haven't been before.
The data is saved to the database.
The data is pulled from the database and displayed with a heatmap UI component on the Patient Dashboard.
The session chosen is currently hardcoded and there is no pausinng the display/changing time.
The heatmap also hasn't been added to pages other than the Patient Dashboard.